pyproject.toml
[project]
name = "strands-aws-bedrock-agent-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=1.28.0",
    "python-dotenv>=1.0.0",
]
